fix(auth): apply the Clerk-load wait to apiFetch too
getSessionToken had the same race as the Supabase client: mount-time
apiFetch calls (dashboard, integrations status, learnings, email
settings, outreach) could read window.Clerk before it loaded and go out
with no Bearer token. The cookie fallback usually covers it, but a
stale __session cookie on a fresh visit means a spurious 401 and an
"expired session" toast on first paint.

The bounded wait now lives in getSessionToken as the single source of
truth; the Supabase client reuses it instead of keeping its own copy.
